updated database to allow for decimals

+++ server/db/models/Item.js
@@ -6,12 +6,12 @@ var db = require('.././_db.js');
 
 var Item = db.define('item', {
   name: Sequelize.STRING,
-  positionX: Sequelize.INTEGER,
-  positionY: Sequelize.INTEGER,
-  positionZ: Sequelize.INTEGER,
-  scaleX: Sequelize.INTEGER,
-  scaleY: Sequelize.INTEGER,
-  scaleZ: Sequelize.INTEGER,
+  positionX: Sequelize.DECIMAL,
+  positionY: Sequelize.DECIMAL,
+  positionZ: Sequelize.DECIMAL,
+  scaleX: Sequelize.DECIMAL,
+  scaleY: Sequelize.DECIMAL,
+  scaleZ: Sequelize.DECIMAL,
   message: Sequelize.TEXT
 });
